refactor(app): fix nominclature and documentation
