+++ bin/cli
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('../src/createTypeAndJS.js')
+require('../src/bootstrap.js')
