+++ index.ts
@@ -1,6 +1,6 @@
 /**
  * Created by tushar on 30/12/17.
  */
-import {createMergedConfig} from './src/createMergedConfig'
+import {mergeAllConfigs} from './src/mergeAllConfigs'
 
-export const config: Config = createMergedConfig(process)
+export const config: Config = mergeAllConfigs(process)
